Backend/requirements.txt
fastapi
uvicorn[standard]
pandas
python-dotenv
google-generativeai
pathway
pillow